+++ tooling/src/main/kotlin/org/jetbrains/spek/tooling/PathType.kt
@@ -1,0 +1,14 @@
+package org.jetbrains.spek.tooling
+
+import com.fasterxml.jackson.annotation.JsonProperty
+
+enum class PathType {
+    @JsonProperty("spec")
+    SPEC,
+
+    @JsonProperty("group")
+    GROUP,
+
+    @JsonProperty("test")
+    TEST
+}